code: refactor kobra.py to be a thin bootstrap entrypoint
Reduced kobra.py from 589 lines to 36 lines by removing all legacy
menu, game state, and drawing functions. All functionality now properly
delegated to ECS architecture (scenes, systems, components).

Changes:
- Removed legacy menu functions (start_menu, run_settings_menu, etc)
- Removed game state drawing functions (draw_grid, draw_pause_screen)
- Removed game over handler and utility functions
- Kept only main() bootstrap that initializes and runs ECSGameApp
- Created kobra_legacy.py with old code for reference

Result: 589 → 36 lines (94% reduction)

Follows ECS project structure: kobra.py is now truly a thin bootstrap
as specified in ecs_project_structure.mdc.
